refactor: update kassandra and kacy token logo
